requirements.txt
aiohttp==3.9.3
aiohttp-socks==0.8.4
loguru==0.7.2
pydantic==2.6.1
pydantic-settings==2.2.1
PyYAML==6.0.1
uvloop==0.19.0 ; sys_platform != 'win32'
